src/test/java/com/leetcode/tree/medium/DeepestLeavesSum_1302Test.java
package com.leetcode.tree.medium;

import com.leetcode.utils.BinaryTree;
import com.leetcode.utils.BinaryTree.TreeNode;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class DeepestLeavesSum_1302Test {

  private final DeepestLeavesSum_1302 solution = new DeepestLeavesSum_1302();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(
          arguments(BinaryTree.of(1, 2, 3, 4, 5, null, 6, 7, null, null, null, null, 8), 15));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void deepestLeavesSum(TreeNode preorder, int expectedResult) {
    int actualResult = solution.deepestLeavesSum(preorder);

    assertThat(actualResult).isEqualTo(expectedResult);
  }
}
